Folds.jl

Folds.Folds — Module

Folds: sequential, threaded, and distributed fold interface for Julia

NOTE: For information on the released version, please see https://juliafolds.github.io/Folds.jl/stable

Folds.jl provides a unified interface for sequential, threaded, and distributed folds.

julia> using Folds

julia> Folds.sum(1:10)
55

julia> Folds.sum(1:10, ThreadedEx())  # equivalent to above
55

julia> Folds.sum(1:10, DistributedEx())
55

Folds.collect — Function

Folds.collect(collection; [init] [executor_options...]) :: AbstractArray
Folds.collect(collection, executor; [init]) :: AbstractArray

Folds.map — Function

Folds.map(f, collections...; [executor_options...])
Folds.map(f, collections..., executor)

Folds.mapreduce — Function

Folds.mapreduce(f, op, collections...; [init] [executor_options...])
Folds.mapreduce(f, op, collections..., executor; [init])

Folds.reduce — Function

Folds.reduce(op, collection; [init] [executor_options...])
Folds.reduce(op, collection, executor; [init])

Folds.sum — Function

Folds.sum([f,] collection; [init] [executor_options...])
Folds.sum([f,] collection, executor; [init])