Folds.jl

Folds — Module

Folds: sequential, threaded, and distributed fold interface for Julia

Folds.jl provides a unified interface for sequential, threaded, and distributed folds.

julia> using Folds

julia> fold(+, 1:10; threaded = true)
55

julia> sum(1:10, Folds.Scheduler(distributed = true))
55

Folds.fold — Function

fold(rf, [xf,] collection; [init], scheduler_options...)
fold(rf, [xf,] collection, scheduler; [init])

Generic entry point for various fold implementations.

See Folds.Scheduler for how to construct scheduler. Keyword arguments scheduler_options are passed Folds.Scheduler.

Keyword Arguments

init: An initial value fed to the first argument to reducing step function step.

sequential::Union{Val{true}, Val{false}, Bool} = false: Whether or not to use left fold.

distributed::Union{Val{true}, Val{false}, Bool} = false: Whether or not to use Distributed.jl.

threaded::Union{Val{true}, Val{false}, Bool} = !sequential: Whether or not to use multi-threading.

ordered::Union{Val{true}, Val{false}, Bool} = true: Whether or not to respect order; if false, assume that the reducing step function is commutative (and associative).

simd::Union{Val{true}, Val{false}, Val{:ivdep}, Bool, Symbol} = false: Whether or not to use SIMD.

Examples

julia> using Folds

julia> xs = 1:10;

julia> fold(+, xs)  # DWIM (threaded)
55

julia> fold(+, xs; sequential = true, simd = true)
55

julia> fold(+, xs; threaded = true)
55

julia> fold(+, xs; distributed = true)  # `threaded = true` implied
55

julia> fold(+, xs; distributed = true, threaded = false)
55

julia> fold(+, xs, Folds.Scheduler(distributed = true, threaded = false))  # equivalent
55

Folds.Scheduler — Type

Folds.Scheduler(; sequential, threaded, distributed, ordered, simd, kwargs...)

Schedule how folds are executed. See fold.

Currently, the default is threaded = true. It is planned to detect parallelizability by assuming the purity of user-defined functions. Pass sequential = true or use Folds.foldl for sequential fold.

Keyword Arguments

sequential::Union{Val{true}, Val{false}, Bool} = false: Whether or not to use left fold.

distributed::Union{Val{true}, Val{false}, Bool} = false: Whether or not to use Distributed.jl.

threaded::Union{Val{true}, Val{false}, Bool} = !sequential: Whether or not to use multi-threading.

ordered::Union{Val{true}, Val{false}, Bool} = true: Whether or not to respect order; if false, assume that the reducing step function is commutative (and associative).

simd::Union{Val{true}, Val{false}, Val{:ivdep}, Bool, Symbol} = false: Whether or not to use SIMD.

Folds.foldl — Function

Folds.foldl(rf, [xf,] collection; [init, simd])
Folds.reduce(rf, [xf,] collection; [init, simd, basesize])

Left fold and threaded reduce based on Transducers.jl.

Arguments

rf: A reducing step function (acc, x) -> acc′.

xf::Transducer: A transducer.

collection: An array, dictionary, iterator, etc.

init: An initial value fed to the first argument to reducing step function step. This argument can be omitted for well know binary operations like + or *. Supported binary operations are listed in InitialValues.jl documentation.

simd: If true or :ivdep, enable SIMD using Base.@simd. If :ivdep, use @simd ivdep for ... end variant. Read Julia manual of Base.@simd to understand when it is appropriate to use this option. This option has no effect if false (default).

basesize::Integer = amount(reducible) ÷ nthreads(): A size of chunk in reducible that is processed by each worker. A smaller size may be required when:

computation time for processing each item fluctuates a lot

computation can be terminated by reduced API.